+++ cmd/sointu-track/main.go
@@ -47,7 +47,10 @@ func main() {
 	midiContext := gomidi.NewContext()
 	defer midiContext.Close()
 	midiContext.TryToOpenBy(*defaultMidiInput, *firstMidiInput)
-	model, player := tracker.NewModelPlayer(cmd.MainSynther, midiContext, recoveryFile)
+	broker := tracker.NewBroker()
+	model, player := tracker.NewModelPlayer(broker, cmd.MainSynther, midiContext, recoveryFile)
+	detector := tracker.NewDetector(broker)
+	go detector.Run()
 
 	if a := flag.Args(); len(a) > 0 {
 		f, err := os.Open(a[0])